Move all ll flags into ll::flags
Alternative to #551
