+++ src/lib.rs
@@ -35,20 +35,20 @@ pub use crate::ll::Errno;
 pub use crate::ll::Generation;
 pub use crate::ll::RequestId;
 pub use crate::ll::TimeOrNow;
-pub use crate::ll::copy_file_range_flags::CopyFileRangeFlags;
-pub use crate::ll::fuse_abi::FopenFlags;
-pub use crate::ll::fuse_abi::InitFlags;
+pub use crate::ll::flags::CopyFileRangeFlags;
+pub use crate::ll::flags::FopenFlags;
+pub use crate::ll::flags::InitFlags;
+pub use crate::ll::flags::IoctlFlags;
+pub use crate::ll::flags::PollFlags;
+pub use crate::ll::flags::ReadFlags;
+pub use crate::ll::flags::WriteFlags;
 pub use crate::ll::fuse_abi::consts;
 use crate::ll::fuse_abi::consts::*;
 pub use crate::ll::fuse_abi::fuse_forget_one;
-pub use crate::ll::ioctl_flags::IoctlFlags;
-pub use crate::ll::poll_flags::PollFlags;
-pub use crate::ll::read_flags::ReadFlags;
 pub use crate::ll::request::FileHandle;
 pub use crate::ll::request::INodeNo;
 pub use crate::ll::request::LockOwner;
 pub use crate::ll::request::Version;
-pub use crate::ll::write_flags::WriteFlags;
 pub use crate::mnt::mount_options::MountOption;
 use crate::mnt::mount_options::check_option_conflicts;
 use crate::mnt::mount_options::parse_options_from_args;
