--- src/ll/copy_file_range_flags.rs
@@ -1,7 +1,0 @@
-use bitflags::bitflags;
-
-bitflags! {
-    /// Flags of `copy_file_range`.
-    #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-    pub struct CopyFileRangeFlags: u64 {}
-}
