+++ src/ll/mod.rs
@@ -1,18 +1,11 @@
 //! Low-level kernel communication.
 
 mod argument;
-pub(crate) mod copy_file_range_flags;
-pub(crate) mod cuse_init_flags;
-pub(crate) mod fsync_flags;
+pub(crate) mod flags;
 pub(crate) mod fuse_abi;
-pub(crate) mod ioctl_flags;
 pub(crate) mod notify;
-pub(crate) mod poll_flags;
-pub(crate) mod read_flags;
-pub(crate) mod release_flags;
 pub(crate) mod reply;
 pub(crate) mod request;
-pub(crate) mod write_flags;
 
 use std::convert::TryInto;
 use std::num::NonZeroI32;
